[MISC][RL] Extract time picker dropdown
